feat(ui): TE-1072 moved SelectDataset to components folder
