+++ thirdeye-ui/src/app/components/welcome-onboard-datasource/wizard-bottom-bar/wizard-bottom-bar.interfaces.ts
